+++ code/tests/functional/tests/backend_api/default/test_azure_byod.py
@@ -4,11 +4,11 @@ from pytest_httpserver import HTTPServer
 import requests
 from string import Template
 
-from tests.functional.backend_api.request_matching import (
+from tests.functional.request_matching import (
     RequestMatcher,
     verify_request_made,
 )
-from tests.functional.backend_api.app_config import AppConfig
+from tests.functional.app_config import AppConfig
 
 pytestmark = pytest.mark.functional
