+++ code/tests/functional/tests/backend_api/sk_orchestrator/conftest.py
@@ -1,7 +1,7 @@
 import logging
 import pytest
-from tests.functional.backend_api.app_config import AppConfig
-from tests.functional.backend_api.common import get_free_port, start_app
+from tests.functional.app_config import AppConfig
+from tests.functional.tests.backend_api.common import get_free_port, start_app
 from backend.batch.utilities.helpers.config.config_helper import ConfigHelper
 from backend.batch.utilities.helpers.env_helper import EnvHelper
 
@@ -29,6 +29,7 @@ def app_config(make_httpserver, ca):
                 "AZURE_SEARCH_SERVICE": f"https://localhost:{make_httpserver.port}/",
                 "AZURE_CONTENT_SAFETY_ENDPOINT": f"https://localhost:{make_httpserver.port}/",
                 "AZURE_SPEECH_REGION_ENDPOINT": f"https://localhost:{make_httpserver.port}/",
+                "ORCHESTRATION_STRATEGY": "semantic_kernel",
                 "SSL_CERT_FILE": ca_temp_path,
                 "CURL_CA_BUNDLE": ca_temp_path,
             }
